+++ django_occupations/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import ONetAlternateTitle, ONetOccupation, SOCDetailedOccupation, SOCBroadOccupation, SOCMinorGroup, SOCMajorGroup, SOCIntermediateAggregationGroup, SOCHighLevelAggregationGroup, SOCDirectMatchTitles
+from .models import ONetAlternateTitle, ONetOccupation, SOCDetailedOccupation, SOCBroadOccupation, SOCMinorGroup, SOCMajorGroup, SOCIntermediateAggregationGroup, SOCHighLevelAggregationGroup
 
 admin.site.register(ONetAlternateTitle)
 admin.site.register(ONetOccupation)
@@ -9,4 +9,3 @@ admin.site.register(SOCMinorGroup)
 admin.site.register(SOCMajorGroup)
 admin.site.register(SOCIntermediateAggregationGroup)
 admin.site.register(SOCHighLevelAggregationGroup)
-admin.site.register(SOCDirectMatchTitles)